feat: implement notification system with WebSocket support and context management
